Support for apostrophe in words
If a quote is written in a word it is handled as a word instead of
starting a string. This fixes the following YAML syntax highlighting:

  server's:
    - here: syntax is not string
    - this: 'is a string'

Previously the first apostrophe/quote started a string, which meant
that the file is highlighted as a string until the next quote is met.

The single quote string detection now requires that the previous character
is a whitespace, which means that:

  this:'is detected incorrectly'

without the space the above does not detect a string.