fill while in a comment should behave as in text modes
This fixes a bug in fill-paragraph that made adaptive fill not work
while in a comment (probably the only time it matters in yaml-mode).

This change does what lisp-mode does: first, try fill-comment-paragraph.
If fill-comment-paragraph does the fill, great; we're in a comment, and
the right thing happens.  If it returns nil, it can't do the fill, and
we can fall back to the previous method.

Without this change, a paragraph break inside a comment, or other things
that adaptive fill would generally detect, are not respected.