Do not modify the buffer in `yaml-indent-line` unless necessary
Calling (yaml-indent-line) for the first time on a line that the
function already deems indented correctly results in the buffer
getting modified even though there was nothing to modify. This is
because the function unconditionally removes space at the beginning of
line and then re-indents text back.

Fix this by replacing the `bol` + `delete-horizontal-space` + `indent-to`
combination with a call to `indent-line-to` which does the right
thing.